requirements.txt
fastapi
uvicorn
motor
pymongo
redis
cachetools
pydantic
python-dotenv
